Add export functionality for exporting content types in content creation format.
The ``ftw.inflator.export:default`` profile adds an entry to the Plone control panel,
allowing to export the current content as ZIP with JSON-files.
The ``content_creation`` folder in the ZIP can be placed in a Generic Setup profile,
which will then create the content when installed.

The export may be incomplete, especially for types with additional features, such as
storing data in the annotations.